Basic registration summary update

+++ validator/src/main/java/com/google/daq/mqtt/registrar/Registrar.java
@@ -1,7 +1,9 @@
 package com.google.daq.mqtt.registrar;
 
+import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.collect.Sets.intersection;
 import static com.google.udmi.util.Common.NO_SITE;
+import static com.google.udmi.util.GeneralUtils.ifNotNullGet;
 import static com.google.udmi.util.JsonUtil.OBJECT_MAPPER;
 import static java.util.Optional.ofNullable;
 
@@ -21,7 +23,6 @@ import com.google.daq.mqtt.util.DeviceExceptionManager;
 import com.google.daq.mqtt.util.ExceptionMap;
 import com.google.daq.mqtt.util.ExceptionMap.ErrorTree;
 import com.google.daq.mqtt.util.PubSubPusher;
-import com.google.udmi.util.Common;
 import com.google.udmi.util.SiteModel;
 import java.io.File;
 import java.io.FileInputStream;
@@ -47,12 +48,14 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
+import org.jetbrains.annotations.NotNull;
 import udmi.schema.CloudModel;
 import udmi.schema.Credential;
 import udmi.schema.Envelope.SubFolder;
 import udmi.schema.ExecutionConfiguration;
 import udmi.schema.ExecutionConfiguration.IotProvider;
 import udmi.schema.Metadata;
+import udmi.schema.SetupUdmiConfig;
 
 /**
  * Validate devices' static metadata and register them in the cloud.
@@ -71,7 +74,7 @@ public class Registrar {
   private static final String VERSION_KEY = "Version";
   private static final String VERSION_MAIN_KEY = "main";
   private static final String SCHEMA_SUFFIX = ".json";
-  private static final String REGISTRATION_SUMMARY_JSON = "registration_summary.json";
+  private static final String REGISTRATION_SUMMARY_JSON = "out/registration_summary.json";
   private static final String SCHEMA_NAME = "UDMI";
   private static final String SITE_METADATA_JSON = "site_metadata.json";
   private static final String SWARM_SUBFOLDER = "swarm";
@@ -94,7 +97,7 @@ public class Registrar {
   private Duration idleLimit;
   private Set<String> cloudDevices;
   private Metadata siteMetadata;
-  private Map<String, Map<String, String>> lastErrorSummary;
+  private Map<String, Object> lastErrorSummary;
   private boolean validateMetadata = false;
   private List<String> deviceList;
   private Boolean blockUnknown;
@@ -113,6 +116,18 @@ public class Registrar {
   public static void main(String[] args) {
     ArrayList<String> argList = new ArrayList<>(List.of(args));
     new Registrar().processArgs(argList).execute();
+  }
+
+  @SuppressWarnings("unchecked")
+  @NotNull
+  private static Map<String, String> getErrorKeyMap(Map<String, Object> resultMap,
+      String errorKey) {
+    return (Map<String, String>) resultMap.computeIfAbsent(errorKey, cat -> new TreeMap<>());
+  }
+
+  @SuppressWarnings("unchecked")
+  private static int getErrorSummaryDetail(Object value) {
+    return ((Map<String, Object>) value).size();
   }
 
   Registrar processArgs(List<String> argListRaw) {
@@ -167,7 +182,7 @@ public class Registrar {
   }
 
   private void setDeleteDevices(boolean deleteDevices) {
-    Preconditions.checkNotNull(projectId, "delete devices specified with no target project");
+    checkNotNull(projectId, "delete devices specified with no target project");
     this.deleteDevices = deleteDevices;
   }
 
@@ -241,7 +256,7 @@ public class Registrar {
     }
   }
 
-  protected Map<String, Map<String, String>> getLastErrorSummary() {
+  protected Map<String, Object> getLastErrorSummary() {
     return lastErrorSummary;
   }
 
@@ -250,7 +265,7 @@ public class Registrar {
       return;
     }
 
-    Map<String, Map<String, String>> errorSummary = new TreeMap<>();
+    Map<String, Object> errorSummary = new TreeMap<>();
     DeviceExceptionManager dem = new DeviceExceptionManager(siteDir);
     localDevices
         .values()
@@ -261,16 +276,14 @@ public class Registrar {
             device -> {
               Set<Entry<String, ErrorTree>> entries =
                   device.getTreeChildren(dem.forDevice(device.getDeviceId()));
-              entries
-                  .forEach(
-                      error ->
-                          errorSummary
-                              .computeIfAbsent(error.getKey(), cat -> new TreeMap<>())
-                              .put(device.getDeviceId(), error.getValue().message));
               if (entries.isEmpty()) {
-                errorSummary
-                    .computeIfAbsent("Clean", cat -> new TreeMap<>())
+                getErrorKeyMap(errorSummary, "Clean")
                     .put(device.getDeviceId(), device.getNormalizedTimestamp());
+              } else {
+                entries
+                    .forEach(
+                        error -> getErrorKeyMap(errorSummary, error.getKey())
+                            .put(device.getDeviceId(), error.getValue().message));
               }
             });
     if (blockErrors != null && !blockErrors.isEmpty()) {
@@ -280,19 +293,28 @@ public class Registrar {
               .collect(Collectors.toMap(Map.Entry::getKey, entry -> entry.getValue().toString())));
     }
     System.err.println("\nSummary:");
-    errorSummary.forEach(
-        (key, value) -> System.err.println("  Device " + key + ": " + value.size()));
+    errorSummary.forEach((key, value) -> System.err.println(
+            "  Device " + key + ": " + getErrorSummaryDetail(value)));
     System.err.println("Out of " + localDevices.size() + " total.");
-    String version = Common.getUdmiVersion();
-    errorSummary.put(VERSION_KEY, Map.of(VERSION_MAIN_KEY, version));
+    errorSummary.put(VERSION_KEY, versionInformation());
     OBJECT_MAPPER.writeValue(summaryFile, errorSummary);
     lastErrorSummary = errorSummary;
+    System.err.println("Registration summary available in " + summaryFile.getAbsolutePath());
+  }
+
+  private SetupUdmiConfig versionInformation() {
+    return ifNotNullGet(cloudIotManager, CloudIotManager::getVersionInformation,
+        CloudIotManager.getDefaultVersion());
   }
 
   protected void setSitePath(String sitePath) {
-    Preconditions.checkNotNull(SCHEMA_NAME, "schemaName not set yet");
+    checkNotNull(SCHEMA_NAME, "schemaName not set yet");
     siteDir = new File(sitePath);
     summaryFile = new File(siteDir, REGISTRATION_SUMMARY_JSON);
+    File parentFile = summaryFile.getParentFile();
+    if (!parentFile.isDirectory() && !parentFile.mkdirs()) {
+      throw new IllegalStateException("Could not create directory " + parentFile.getAbsolutePath());
+    }
     summaryFile.delete();
   }
 
@@ -486,9 +508,9 @@ public class Registrar {
 
     updateCloudIoT(localDevice);
     CloudModel device =
-        Preconditions.checkNotNull(fetchDevice(localName), "missing device " + localName);
+        checkNotNull(fetchDevice(localName), "missing device " + localName);
     String numId =
-        Preconditions.checkNotNull(
+        checkNotNull(
             device.num_id, "missing deviceNumId for " + localName);
     localDevice.setDeviceNumId(numId);
   }
@@ -649,7 +671,7 @@ public class Registrar {
   }
 
   private Map<String, LocalDevice> loadLocalDevices(Set<String> specifiedDevices) {
-    Preconditions.checkNotNull(siteDir, "site directory");
+    checkNotNull(siteDir, "site directory");
     File devicesDir = new File(siteDir, DEVICES_DIR);
     if (!devicesDir.isDirectory()) {
       throw new RuntimeException("Not a valid directory: " + devicesDir.getAbsolutePath());
